pyproject.toml
[tool.poetry]
name = "enhanced-file-merger"
version = "0.1.0"
description = "A tool to merge source code files with metadata"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "merge_files", from = "src"}, {include = "tests"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^24.1.0"
isort = "^5.13.2"
mypy = "^1.8.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
merge-files = "merge_files.merge_files:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--verbose --cov=merge_files --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"